POS Store API: move fee routes into POS, keep only the store shared
Address the public-placement concern: a fee-write endpoint is POS-only, so the
route classes should not sit in the public Store API routes directory. Move the
cart/add-fee and cart/remove-fee logic into the POS routes (extending
AbstractCartRoute directly + PosRouteTrait) and delete the StoreApi\Routes\V1
base classes. The reusable persistence engine (CustomFeesStore) stays shared in
StoreApi\Utilities, where cross-consumer reuse is real. Behaviour is unchanged.
